--- DesktopEdge/Views/Screens/MFAAuthScreen.xaml
@@ -1,12 +1,0 @@
-﻿<UserControl x:Class="Ziti.Desktop.Edge.Views.Screens.MFAAuthScreen"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:Ziti.Desktop.Edge.Views.Screens"
-             mc:Ignorable="d" 
-             d:DesignHeight="450" d:DesignWidth="800">
-    <Grid>
-            
-    </Grid>
-</UserControl>
