+++ DesktopEdge/Views/Screens/MFAScreen.xaml
@@ -31,7 +31,7 @@
             <Grid.RowDefinitions>
                 <RowDefinition Height="30"></RowDefinition>
                 <RowDefinition Height="30"></RowDefinition>
-                <RowDefinition Height="245"></RowDefinition>
+                <RowDefinition Height="200"></RowDefinition>
                 <RowDefinition Height="30"></RowDefinition>
                 <RowDefinition Height="30"></RowDefinition>
                 <RowDefinition Height="30"></RowDefinition>
@@ -41,8 +41,13 @@
             </Grid.RowDefinitions>
             <Label Content="Setup MFA" Grid.Row="0" FontSize="16" HorizontalAlignment="Center" FontWeight="Bold"></Label>
             <Label x:Name="IdName" Content="Identity Name" Grid.Row="1" FontSize="12" Foreground="DarkGray" HorizontalAlignment="Center"></Label>
-            <Image x:Name="MFAImage" MaxWidth="200" MaxHeight="200" Width="200" Height="200" RenderOptions.BitmapScalingMode="Fant" Stretch="Fill" Source="/Assets/Images/qrcode.png" Margin="50,19,50,0" Grid.RowSpan="4" />
-            <Label Grid.Row="3" HorizontalAlignment="Center" Content="Go To Link" Foreground="#0069FF" MouseUp="GoTo" Cursor="Hand"></Label>
+            <Image x:Name="MFAImage" MaxWidth="200" MaxHeight="200" Width="200" Height="200" RenderOptions.BitmapScalingMode="Fant" Stretch="Fill" Source="/Assets/Images/qrcode.png" Margin="50,19,50,0" Grid.Row="2" />
+            <TextBlock x:Name="SecretCode" TextWrapping="WrapWithOverflow" Grid.Row="2" HorizontalAlignment="Center" VerticalAlignment="Center"></TextBlock>
+            <StackPanel Orientation="Horizontal" HorizontalAlignment="Center" Grid.Row="3" Grid.Column="1" VerticalAlignment="Center">
+                <Label HorizontalAlignment="Center" Content="Go To Link" Foreground="#0069FF" MouseUp="GoTo" Cursor="Hand"></Label>
+                <Label HorizontalAlignment="Center" Content=" or " Foreground="#000000" Cursor="Hand"></Label>
+                <Label x:Name="SecretButton" HorizontalAlignment="Center" Content="Show Secret" Foreground="#0069FF" MouseUp="ShowSecret" Cursor="Hand"></Label>
+            </StackPanel>
             <Label Content="enter the code below to finish setup" Grid.Row="4" FontSize="11" Foreground="DarkGray" HorizontalAlignment="Center"></Label>
             <Label Content="Authentication Code" Grid.Row="5" FontSize="18" Foreground="#FFFFFF" HorizontalAlignment="Center" FontWeight="Bold" FontFamily="/ZitiDesktopEdge;component/Assets/Fonts/#Russo One"></Label>
             <Grid Grid.Row="6" HorizontalAlignment="Center">
